refactor(mpp): address PR #181 structural review (confidential transfers)
Behavior-preserving structural changes from the review; verified by the
full lib suite + confidential litesvm e2e (737 pass) and clippy/fmt.

- H1: extract the confidential subsystem (settle_confidential_bundle,
  sweep_confidential_orphans, broadcast_close, verify_confidential_bundle_tx,
  orphan guard, consts, ConfidentialSweepReport) into server/confidential.rs
  as `impl Mpp`. Mpp fields promoted to pub(crate); shared helpers
  (resolve_expected_mint, reject_address_lookup_tables, consume_signature,
  the ComputeBudget consts) made pub(crate). ~950 lines leave charge.rs.
- H2: dedup the ComputeBudget decoder — extract decode_compute_budget_op /
  ComputeBudgetOp; each call site keeps its own caps + error type, so the
  plaintext and confidential paths stay byte-for-byte behavior-identical.
- H5: collapse the three copy-paste proof-context blocks in the client
  bundle builder into proof_context_pair<T>, returning (Keypair, [create,
  verify]) so the range site still feeds the record-staging path unchanged.
- H4: kill the triplicated POD casts — client::confidential is now the one
  canonical (pub(crate)) copy; the protocol litesvm test calls it instead of
  re-deriving (no more "Same as client/confidential.rs" drift hazard).
- Docs: confidential-transfers.md now reflects the upstream litesvm patch
  (LiteSVM/litesvm@master, not the fork) and drops the nonexistent
  RUN_SURFPOOL_TESTS gate (the e2e tests live behind the `confidential`
  Cargo feature, which CI does not enable).

Build fix: solana-keypair 3.1 pulls five8 1.0.0 -> five8_core 0.1.2, whose
Error impl is std-gated; pull five8_core with `std` so the build resolves.
