+++ rust/crates/mpp/src/client/mod.rs
@@ -3,7 +3,7 @@
 pub mod authenticate;
 mod charge;
 #[cfg(feature = "confidential")]
-mod confidential;
+pub(crate) mod confidential;
 #[cfg(feature = "confidential")]
 pub use confidential::{build_confidential_transfer_bundle, ConfidentialTransferParams};
 pub mod http_stream;
